refactor: 중복 검증 역할을 save에서 분리
