+++ src/main/java/wooteco/subway/service/StationService.java
@@ -6,6 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
+import wooteco.subway.exception.StationDuplicateException;
 import wooteco.subway.exception.StationNotFoundException;
 
 @Service
@@ -19,7 +20,14 @@ public class StationService {
 
     public Station create(final StationRequest stationRequest) {
         final Station station = stationRequest.toEntity();
+        checkDuplicateName(station);
         return stationDao.save(station);
+    }
+
+    private void checkDuplicateName(final Station station) {
+        if (stationDao.existsName(station)) {
+            throw new StationDuplicateException("이미 존재하는 지하철역 이름입니다.");
+        }
     }
 
     @Transactional(readOnly = true)
